Refactor and update README.md
