+++ src/main/scala/com/creatorships/web/comic/catalogs/infrastructure/DistributorsRepository.scala
@@ -4,7 +4,7 @@ import cats.Monad
 import cats.effect.Bracket
 import com.creatorships.web.comic.catalogs.application.distributors.{Coordinators, Publishers}
 import com.creatorships.web.comic.catalogs.domain.provider.Provider.{Coordinator, Publisher}
-import com.creatorships.web.comic.catalogs.infrastructure.analyze.url.DistributorType
+import com.creatorships.web.comic.catalogs.infrastructure.analyze.rule.DistributorType
 import doobie.implicits._
 import doobie.util.fragment.Fragment
 import doobie.util.transactor.Transactor
